+++ src/i2c.rs
@@ -10,6 +10,7 @@ use crate::rcc::Clocks;
 use crate::time::{Hertz, U32Ext};
 
 mod hal_02;
+mod hal_1;
 
 #[derive(Debug, Eq, PartialEq)]
 pub enum DutyCycle {
@@ -98,16 +99,37 @@ where
     }
 }
 
+pub use embedded_hal_one::i2c::NoAcknowledgeSource;
+
 #[derive(Debug, Eq, PartialEq, Copy, Clone)]
+#[non_exhaustive]
 pub enum Error {
-    OVERRUN,
-    NACK,
-    TIMEOUT,
-    // Note: The BUS error type is not currently returned, but is maintained for backwards
-    // compatibility.
-    BUS,
-    CRC,
-    ARBITRATION,
+    Overrun,
+    NoAcknowledge(NoAcknowledgeSource),
+    Timeout,
+    // Note: The Bus error type is not currently returned, but is maintained for compatibility.
+    Bus,
+    Crc,
+    ArbitrationLoss,
+}
+
+impl Error {
+    pub(crate) fn nack_addr(self) -> Self {
+        match self {
+            Error::NoAcknowledge(NoAcknowledgeSource::Unknown) => {
+                Error::NoAcknowledge(NoAcknowledgeSource::Address)
+            }
+            e => e,
+        }
+    }
+    pub(crate) fn nack_data(self) -> Self {
+        match self {
+            Error::NoAcknowledge(NoAcknowledgeSource::Unknown) => {
+                Error::NoAcknowledge(NoAcknowledgeSource::Data)
+            }
+            e => e,
+        }
+    }
 }
 
 pub trait Instance: crate::Sealed + Deref<Target = i2c1::RegisterBlock> + Enable + Reset {}
@@ -224,27 +246,27 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
 
         if sr1.timeout().bit_is_set() {
             self.i2c.sr1.modify(|_, w| w.timeout().clear_bit());
-            return Err(Error::TIMEOUT);
+            return Err(Error::Timeout);
         }
 
         if sr1.pecerr().bit_is_set() {
             self.i2c.sr1.modify(|_, w| w.pecerr().clear_bit());
-            return Err(Error::CRC);
+            return Err(Error::Crc);
         }
 
         if sr1.ovr().bit_is_set() {
             self.i2c.sr1.modify(|_, w| w.ovr().clear_bit());
-            return Err(Error::OVERRUN);
+            return Err(Error::Overrun);
         }
 
         if sr1.af().bit_is_set() {
             self.i2c.sr1.modify(|_, w| w.af().clear_bit());
-            return Err(Error::NACK);
+            return Err(Error::NoAcknowledge(NoAcknowledgeSource::Unknown));
         }
 
         if sr1.arlo().bit_is_set() {
             self.i2c.sr1.modify(|_, w| w.arlo().clear_bit());
-            return Err(Error::ARBITRATION);
+            return Err(Error::ArbitrationLoss);
         }
 
         // The errata indicates that BERR may be incorrectly detected. It recommends ignoring and
@@ -281,7 +303,9 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
         // Wait until address was sent
         loop {
             // Check for any I2C errors. If a NACK occurs, the ADDR bit will never be set.
-            let sr1 = self.check_and_clear_error_flags()?;
+            let sr1 = self
+                .check_and_clear_error_flags()
+                .map_err(Error::nack_addr)?;
 
             // Wait for the address to be acknowledged
             if sr1.addr().bit_is_set() {
@@ -304,14 +328,24 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
     fn send_byte(&self, byte: u8) -> Result<(), Error> {
         // Wait until we're ready for sending
         // Check for any I2C errors. If a NACK occurs, the ADDR bit will never be set.
-        while self.check_and_clear_error_flags()?.tx_e().bit_is_clear() {}
+        while self
+            .check_and_clear_error_flags()
+            .map_err(Error::nack_addr)?
+            .tx_e()
+            .bit_is_clear()
+        {}
 
         // Push out a byte of data
         self.i2c.dr.write(|w| unsafe { w.bits(u32::from(byte)) });
 
         // Wait until byte is transferred
         // Check for any potential error conditions.
-        while self.check_and_clear_error_flags()?.btf().bit_is_clear() {}
+        while self
+            .check_and_clear_error_flags()
+            .map_err(Error::nack_data)?
+            .btf()
+            .bit_is_clear()
+        {}
 
         Ok(())
     }
@@ -319,7 +353,8 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
     fn recv_byte(&self) -> Result<u8, Error> {
         loop {
             // Check for any potential error conditions.
-            self.check_and_clear_error_flags()?;
+            self.check_and_clear_error_flags()
+                .map_err(Error::nack_data)?;
 
             if self.i2c.sr1.read().rx_ne().bit_is_set() {
                 break;
@@ -353,7 +388,8 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
 
             // Wait until address was sent
             loop {
-                self.check_and_clear_error_flags()?;
+                self.check_and_clear_error_flags()
+                    .map_err(Error::nack_addr)?;
                 if self.i2c.sr1.read().addr().bit_is_set() {
                     break;
                 }
@@ -381,7 +417,7 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
             // Fallthrough is success
             Ok(())
         } else {
-            Err(Error::OVERRUN)
+            Err(Error::Overrun)
         }
     }
 
@@ -416,9 +452,7 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
 
     pub fn write_read(&mut self, addr: u8, bytes: &[u8], buffer: &mut [u8]) -> Result<(), Error> {
         self.write_bytes(addr, bytes.iter().cloned())?;
-        self.read(addr, buffer)?;
-
-        Ok(())
+        self.read(addr, buffer)
     }
 
     pub fn write_iter_read<B>(&mut self, addr: u8, bytes: B, buffer: &mut [u8]) -> Result<(), Error>
@@ -426,8 +460,6 @@ impl<I2C: Instance, PINS> I2c<I2C, PINS> {
         B: IntoIterator<Item = u8>,
     {
         self.write_bytes(addr, bytes.into_iter())?;
-        self.read(addr, buffer)?;
-
-        Ok(())
+        self.read(addr, buffer)
     }
 }